Extend variant handling with signals from DBus
The binding tool actually handles quite a few more exotic DBus structures than the standard QDBusAbstractAdaptor can. This causes some of the signals to not correctly connect to QML slots. To get around this we listen for any incomming QML signals and then tell the parent QObject to emit the signal with the correct mapping.

Tested with org.freedesktop.DBus.ObjectManager to get correct signal notifications from UDisks2 to alert to new disks added.

There were some other changes that may break upstream scripts or helpers:

I changed the "test" build target to work with qmake's "make check" feature.  This way I could run the tests quickly and recursively.
I added the qmlplugindump to the build so it can create a plugins.qmltypes.  This is helpful because for some reason Qt Creator's plugin dump process aborted on a bus interface not everything would show up.  This makes it behave a bit more reliable.
I added a "make install" target that will allow for better packaging of the generated bindings.  This will place everything in the correct import directory and name the libraries correctly.
I also changed the Q_PLUGIN_METADATA to the standard Qt interface.

Change-Id: I9917f799a7d76ac091bf67b0e4c2bc3ba1502301